backend/hotel/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path, include
from .views import RoomViewSet, ReservationViewSet

router = DefaultRouter()
router.register(r'rooms', RoomViewSet, basename='room')
router.register(r'reservations', ReservationViewSet, basename='reservation')

urlpatterns = [
    path('', include(router.urls)),
]
