backend/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
django-cors-headers==4.4.0
Pillow==10.3.0
PyMySQL==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
